+++ R/Class-Pop.R
@@ -375,8 +375,6 @@ isNamedMapPop = function(x) {
 #' @slot gxe list containing GxE slopes for GxE traits
 #' @slot fixEff a fixed effect relating to the phenotype.
 #' Used by genomic selection models but otherwise ignored.
-#' @slot reps the number of replications used to measure the
-#' phenotype. Used by genomic selection models, but otherwise ignored.
 #' @slot misc a list whose elements correspond to individuals in the
 #' population. This list is normally empty and exists solely as an
 #' open slot available for uses to store extra information about
@@ -395,7 +393,6 @@ setClass("Pop",
                  ebv="matrix",
                  gxe="list",
                  fixEff="integer",
-                 reps="numeric",
                  misc="list"),
          contains="RawPop")
 
@@ -455,9 +452,6 @@ setValidity("Pop",function(object){
   if(object@nInd!=length(object@fixEff)){
     errors = c(errors,"nInd!=length(fixEff)")
   }
-  if(object@nInd!=length(object@reps)){
-    errors = c(errors,"nInd!=length(reps)")
-  }
   if(object@nInd!=length(object@misc)){
     errors = c(errors,"nInd!=length(misc)")
   }
@@ -490,7 +484,6 @@ setMethod("[",
             x@mother = x@mother[i]
             x@father = x@father[i]
             x@fixEff = x@fixEff[i]
-            x@reps = x@reps[i]
             x@misc = x@misc[i]
             x@gv = x@gv[i,,drop=FALSE]
             x@pheno = x@pheno[i,,drop=FALSE]
@@ -692,7 +685,6 @@ newPop = function(rawPop,simParam=NULL,...){
                mother=mother,
                father=father,
                fixEff=rep(1L,rawPop@nInd),
-               reps=rep(1,rawPop@nInd),
                nTraits=simParam$nTraits,
                gv=gv,
                gxe=gxe,
@@ -764,7 +756,6 @@ resetPop = function(pop,simParam=NULL){
   pop@gv = matrix(NA_real_,nrow=pop@nInd,
                   ncol=simParam$nTraits)
   pop@fixEff = rep(1L,pop@nInd)
-  pop@reps = rep(1,pop@nInd)
   if(simParam$nTraits>=1){
     for(i in 1:simParam$nTraits){
       tmp = getGv(simParam$traits[[i]],pop,simParam$nThreads)
@@ -870,7 +861,6 @@ newEmptyPop = function(ploidy=2L, simParam=NULL){
                mother = character(),
                father = character(),
                fixEff = integer(),
-               reps = integer(),
                nTraits = simParam$nTraits,
                gv = traitMat,
                gxe = vector("list", simParam$nTraits),