+++ src/MME.cpp
@@ -109,15 +109,15 @@ arma::mat makeZ(arma::uvec& z, arma::uword nGeno){
 }
 
 
-
-// Generates weighted matrix
-// Allows for heterogenous variance due to unequal replication
-void sweepReps(arma::mat& X, arma::vec reps){
-  reps = sqrt(reps);
-  for(arma::uword i=0; i<X.n_cols; ++i){
-    X.col(i) = X.col(i)%reps;
-  }
-}
+// Discontinued support
+// // Generates weighted matrix
+// // Allows for heterogenous variance due to unequal replication
+// void sweepReps(arma::mat& X, arma::vec reps){
+//   reps = sqrt(reps);
+//   for(arma::uword i=0; i<X.n_cols; ++i){
+//     X.col(i) = X.col(i)%reps;
+//   }
+// }
 
 //' @title Solve RR-BLUP
 //'
@@ -776,20 +776,20 @@ Rcpp::List callFastRRBLUP(arma::vec y,
 
 // Called by RRBLUP function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP(arma::mat y, arma::uvec x,
                       arma::field<arma::Cube<unsigned char> >& geno, 
                       arma::Col<int>& lociPerChr, arma::uvec lociLoc,
-                      bool useReps, int nThreads){
+                      int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::mat X = makeX(x);
   arma::mat M = genoToGenoA(getGeno(geno,lociPerChr,lociLoc,nThreads),
                             ploidy,nThreads);
   arma::rowvec Mmean = mean(M);
-  if(useReps){
-    sweepReps(y,reps);
-    sweepReps(X,reps);
-    sweepReps(M,reps);
-  }
+  // if(useReps){
+  //   sweepReps(y,reps);
+  //   sweepReps(X,reps);
+  //   sweepReps(M,reps);
+  // }
   Rcpp::List ans = solveRRBLUP(y, X, M);
   arma::vec u = ans["u"];
   arma::mat beta = ans["beta"];
@@ -802,21 +802,21 @@ Rcpp::List callRRBLUP(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP2(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP2(arma::mat y, arma::uvec x, 
                        arma::field<arma::Cube<unsigned char> >& geno, 
                        arma::Col<int>& lociPerChr, arma::uvec lociLoc,
                        double Vu, double Ve, double tol, int maxIter, 
-                       bool useEM, bool useReps, int nThreads){
+                       bool useEM, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::mat X = makeX(x);
   arma::mat M = genoToGenoA(getGeno(geno,lociPerChr,lociLoc,nThreads),
                             ploidy,nThreads);
   arma::rowvec Mmean = mean(M);
-  if(useReps){
-    sweepReps(y,reps);
-    sweepReps(X,reps);
-    sweepReps(M,reps);
-  }
+  // if(useReps){
+  //   sweepReps(y,reps);
+  //   sweepReps(X,reps);
+  //   sweepReps(M,reps);
+  // }
   Rcpp::List ans = solveRRBLUP_EM(y, X, M, Vu, Ve, 
                                   tol, maxIter, useEM);
   arma::vec u = ans["u"];
@@ -830,10 +830,10 @@ Rcpp::List callRRBLUP2(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP_D function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_D(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_D(arma::mat y, arma::uvec x,
                         arma::field<arma::Cube<unsigned char> >& geno, 
                         arma::Col<int>& lociPerChr, arma::uvec lociLoc,
-                        int maxIter, bool useReps, int nThreads){
+                        int maxIter, int nThreads){
   // Fit GS model
   arma::uword ploidy = geno(0).n_cols;
   arma::Mat<unsigned char> M = getGeno(geno,lociPerChr,lociLoc,nThreads);
@@ -843,12 +843,12 @@ Rcpp::List callRRBLUP_D(arma::mat y, arma::uvec x, arma::vec reps,
   Mlist(1) = genoToGenoD(M,ploidy,nThreads);
   arma::mat X;
   X = join_rows(makeX(x),sum(Mlist(1),1));
-  if(useReps){
-    sweepReps(y,reps);
-    sweepReps(X,reps);
-    sweepReps(Mlist(0),reps);
-    sweepReps(Mlist(1),reps);
-  }
+  // if(useReps){
+  //   sweepReps(y,reps);
+  //   sweepReps(X,reps);
+  //   sweepReps(Mlist(0),reps);
+  //   sweepReps(Mlist(1),reps);
+  // }
   Rcpp::List ans = solveRRBLUPMK(y, X, Mlist, maxIter);
   
   // Clear memory
@@ -922,11 +922,11 @@ Rcpp::List callRRBLUP_D(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP_D2 function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_D2(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_D2(arma::mat y, arma::uvec x, 
                          arma::field<arma::Cube<unsigned char> >& geno, 
                          arma::Col<int>& lociPerChr, arma::uvec lociLoc,
                          int maxIter, double Va, double Vd, double Ve, 
-                         double tol, bool useEM, bool useReps, int nThreads){
+                         double tol, bool useEM, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::Mat<unsigned char> M = getGeno(geno,lociPerChr,lociLoc,nThreads);
   arma::mat Ma = genoToGenoA(M,ploidy,nThreads);
@@ -934,12 +934,12 @@ Rcpp::List callRRBLUP_D2(arma::mat y, arma::uvec x, arma::vec reps,
   arma::mat Md = genoToGenoD(M,ploidy,nThreads);
   arma::mat X;
   X = join_rows(makeX(x),sum(Md,1));
-  if(useReps){
-    sweepReps(y,reps);
-    sweepReps(X,reps);
-    sweepReps(Ma,reps);
-    sweepReps(Md,reps);
-  }
+  // if(useReps){
+  //   sweepReps(y,reps);
+  //   sweepReps(X,reps);
+  //   sweepReps(Ma,reps);
+  //   sweepReps(Md,reps);
+  // }
   Rcpp::List ans = solveRRBLUP_EM2(y,X,Ma,Md,Va,Vd,Ve,tol,maxIter,useEM);
   
   // Clear memory
@@ -1014,20 +1014,20 @@ Rcpp::List callRRBLUP_D2(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_MV(arma::mat Y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_MV(arma::mat Y, arma::uvec x,
                          arma::field<arma::Cube<unsigned char> >& geno, 
                          arma::Col<int>& lociPerChr, arma::uvec lociLoc, 
-                         int maxIter, bool useReps, int nThreads){
+                         int maxIter, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::mat X = makeX(x);
   arma::mat M = genoToGenoA(getGeno(geno,lociPerChr,lociLoc,nThreads),
                             ploidy,nThreads);
   arma::rowvec Mmean = mean(M);
-  if(useReps){
-    sweepReps(Y,reps);
-    sweepReps(X,reps);
-    sweepReps(M,reps);
-  }
+  // if(useReps){
+  //   sweepReps(Y,reps);
+  //   sweepReps(X,reps);
+  //   sweepReps(M,reps);
+  // }
   Rcpp::List ans = solveRRBLUPMV(Y, X, M, maxIter);
   arma::mat u = ans["u"];
   arma::mat beta = ans["beta"];
@@ -1040,10 +1040,10 @@ Rcpp::List callRRBLUP_MV(arma::mat Y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP_GCA function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_GCA(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_GCA(arma::mat y, arma::uvec x,
                           arma::field<arma::Cube<unsigned char> >& geno, 
                           arma::Col<int>& lociPerChr, arma::uvec lociLoc, 
-                          int maxIter, bool useReps, int nThreads){
+                          int maxIter, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::field<arma::mat> Mlist(2);
   Mlist(0) = genoToGenoA(getMaternalGeno(geno,lociPerChr,lociLoc,
@@ -1053,12 +1053,12 @@ Rcpp::List callRRBLUP_GCA(arma::mat y, arma::uvec x, arma::vec reps,
   arma::rowvec Mmean1 = mean(Mlist(0));
   arma::rowvec Mmean2 = mean(Mlist(1));
   arma::mat X = makeX(x);
-  if(useReps){
-    sweepReps(y, reps);
-    sweepReps(X, reps);
-    sweepReps(Mlist(0), reps);
-    sweepReps(Mlist(1), reps);
-  }
+  // if(useReps){
+  //   sweepReps(y, reps);
+  //   sweepReps(X, reps);
+  //   sweepReps(Mlist(0), reps);
+  //   sweepReps(Mlist(1), reps);
+  // }
   Rcpp::List ans = solveRRBLUPMK(y,X,Mlist,maxIter);
   arma::field<arma::mat> u = ans["u"];
   arma::mat beta = ans["beta"];
@@ -1073,11 +1073,11 @@ Rcpp::List callRRBLUP_GCA(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP_GCA2 function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_GCA2(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_GCA2(arma::mat y, arma::uvec x, 
                            arma::field<arma::Cube<unsigned char> >& geno, 
                            arma::Col<int>& lociPerChr, arma::uvec lociLoc, 
                            int maxIter, double Vu1, double Vu2, double Ve, 
-                           double tol, bool useEM, bool useReps, int nThreads){
+                           double tol, bool useEM, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::mat M1 = genoToGenoA(getMaternalGeno(geno,lociPerChr,lociLoc,
                                              nThreads),ploidy/2,nThreads);
@@ -1086,12 +1086,12 @@ Rcpp::List callRRBLUP_GCA2(arma::mat y, arma::uvec x, arma::vec reps,
   arma::rowvec Mmean1 = mean(M1);
   arma::rowvec Mmean2 = mean(M2);
   arma::mat X = makeX(x);
-  if(useReps){
-    sweepReps(y, reps);
-    sweepReps(X, reps);
-    sweepReps(M1, reps);
-    sweepReps(M2, reps);
-  }
+  // if(useReps){
+  //   sweepReps(y, reps);
+  //   sweepReps(X, reps);
+  //   sweepReps(M1, reps);
+  //   sweepReps(M2, reps);
+  // }
   Rcpp::List ans = solveRRBLUP_EM2(y,X,M1,M2,Vu1,Vu2,Ve,tol,maxIter,useEM);
   arma::mat u = ans["u"];
   arma::mat beta = ans["beta"];
@@ -1107,10 +1107,10 @@ Rcpp::List callRRBLUP_GCA2(arma::mat y, arma::uvec x, arma::vec reps,
 // Called by RRBLUP_SCA function
 // Only works with diploids
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_SCA(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_SCA(arma::mat y, arma::uvec x, 
                           arma::field<arma::Cube<unsigned char> >& geno, 
                           arma::Col<int>& lociPerChr, arma::uvec lociLoc, 
-                          int maxIter, bool useReps, int nThreads){
+                          int maxIter, int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::field<arma::mat> Mlist(3);
   Mlist(0) = genoToGenoA(getMaternalGeno(geno,lociPerChr,lociLoc,
@@ -1148,13 +1148,13 @@ Rcpp::List callRRBLUP_SCA(arma::mat y, arma::uvec x, arma::vec reps,
   arma::rowvec Mmean12 = mean(Mlist(0)%Mlist(1));
   arma::mat X;
   X = join_rows(makeX(x),sum(Mlist(2),1));
-  if(useReps){
-    sweepReps(y, reps);
-    sweepReps(X, reps);
-    sweepReps(Mlist(0), reps);
-    sweepReps(Mlist(1), reps);
-    sweepReps(Mlist(2), reps);
-  }
+  // if(useReps){
+  //   sweepReps(y, reps);
+  //   sweepReps(X, reps);
+  //   sweepReps(Mlist(0), reps);
+  //   sweepReps(Mlist(1), reps);
+  //   sweepReps(Mlist(2), reps);
+  // }
   
   Rcpp::List ans = solveRRBLUPMK(y, X, Mlist, maxIter);
   
@@ -1228,11 +1228,11 @@ Rcpp::List callRRBLUP_SCA(arma::mat y, arma::uvec x, arma::vec reps,
 
 // Called by RRBLUP_SCA2 function
 // [[Rcpp::export]]
-Rcpp::List callRRBLUP_SCA2(arma::mat y, arma::uvec x, arma::vec reps,
+Rcpp::List callRRBLUP_SCA2(arma::mat y, arma::uvec x, 
                            arma::field<arma::Cube<unsigned char> >& geno, 
                            arma::Col<int>& lociPerChr, arma::uvec lociLoc, 
                            int maxIter, double Vu1, double Vu2, double Vu3, 
-                           double Ve, double tol, bool useEM, bool useReps, 
+                           double Ve, double tol, bool useEM,  
                            int nThreads){
   arma::uword ploidy = geno(0).n_cols;
   arma::mat M1 = genoToGenoA(getMaternalGeno(geno,lociPerChr,lociLoc,
@@ -1270,13 +1270,13 @@ Rcpp::List callRRBLUP_SCA2(arma::mat y, arma::uvec x, arma::vec reps,
   arma::rowvec Mmean12 = mean(M1%M2);
   arma::mat X;
   X = join_rows(makeX(x),sum(M3,1));
-  if(useReps){
-    sweepReps(y, reps);
-    sweepReps(X, reps);
-    sweepReps(M1, reps);
-    sweepReps(M2, reps);
-    sweepReps(M3, reps);
-  }
+  // if(useReps){
+  //   sweepReps(y, reps);
+  //   sweepReps(X, reps);
+  //   sweepReps(M1, reps);
+  //   sweepReps(M2, reps);
+  //   sweepReps(M3, reps);
+  // }
   
   Rcpp::List ans = solveRRBLUP_EM3(y,X,M1,M2,M3,Vu1,Vu2,Vu3,Ve,tol,maxIter,useEM);